# server/internal/handlers/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat App API",
    "version": "0.9.3",
    "description": "实时聊天应用后端 API。\n\n认证方式:\n- `Authorization: Bearer <access_token>` 请求头（首选）\n- `access_token` Cookie（SSE 等 WebSocket 场景回退）\n- `refresh_token` Cookie（仅 `/api/auth/refresh`）\n\n所有错误响应统一格式:\n```json\n{ \"error\": \"error_code\", \"message\": \"human-readable message\" }\n```\n"
  },
  "servers": [
    {
      "url": "https://wsl-8080.moonchan.xyz",
      "description": "生产服务器"
    },
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
  "paths": {
...
    "/api/chats/{chatID}/messages": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ChatID"
        }
      ],
      "get": {
        "summary": "获取消息列表（游标分页）",
        "operationId": "listMessages",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 100
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "游标（消息 ID），获取此消息之前的更早消息"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messages": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "summary": "发送消息",
        "operationId": "sendMessage",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/ContentTooLong"
          }
        }
      }
    },
    "/api/chats/{chatID}/messages/{messageID}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ChatID"
        }
      ],
      "patch": {
        "summary": "编辑消息",
        "operationId": "editMessage",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "summary": "删除消息",
        "operationId": "deleteMessage",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/chats/{chatID}/read": {
      "post": {
        "summary": "标记消息已读",
        "operationId": "markRead",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChatID"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/chats/{chatID}/messages/{messageID}/stream": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ChatID"
        },
        {
          "name": "messageID",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "获取 Stream 消息的实时内容（SSE）",
        "description": "用于 type=stream 的消息，返回 SSE 事件流。先读取内存中的 chunks，结束后 fallback 到 DB。",
        "operationId": "streamMessageContent",
        "tags": [
          "Messages"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE 事件流",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "`Authorization: Bearer <access_token>`\naccess_token 有效期 15 分钟，通过 `/api/auth/refresh` 续期。\n"
      }
    },
...
    "parameters": {
      "ChatID": {
        "name": "chatID",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "聊天 ID"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "bad_request",
                    "message": "invalid request body"
                  }
                }
              ]
            }
          }
        }
      },
...
      "Forbidden": {
        "description": "无权限",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "forbidden",
                    "message": "you are not a member of this chat"
                  }
                }
              ]
            }
          }
        }
      },
...
      "ContentTooLong": {
        "description": "内容超长",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "content_too_long",
                    "message": "content exceeds maximum length"
                  }
                }
              ]
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码（稳定，可用于客户端判断）",
            "example": "bad_request"
          },
          "message": {
            "type": "string",
            "description": "人类可读的错误描述",
            "example": "invalid request body"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "avatar_color",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123def456"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "alice"
          },
          "avatar_color": {
            "type": "string",
            "description": "默认头像背景色（Hex）",
            "example": "#5865F2"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri",
            "example": "https://upload.moonchan.xyz/api/xxx/avatar.png"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "away"
            ],
            "example": "online"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AISource": {
        "type": "object",
        "required": [
          "endpoint",
          "auth_key",
          "body"
        ],
        "properties": {
          "endpoint": {
            "type": "string",
            "format": "uri",
            "description": "AI 上游 API URL，如 https://api.siliconflow.cn/v1/chat/completions"
          },
          "auth_key": {
            "type": "string",
            "description": "Authorization Bearer token"
          },
          "body": {
            "type": "object",
            "description": "请求体 JSON，stream:true 会被自动注入"
          }
        }
      },
      "Message": {
        "type": "object",
        "required": [
          "id",
          "chat_id",
          "user_id",
          "content",
          "created_at",
          "attachment_count",
          "mention_count",
          "reaction_count"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "chat_id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "消息类型。空字符串为普通消息，\"stream\" 为 AI 流式消息。"
          },
          "stream_url": {
            "type": "string",
            "description": "type=stream 时存在，SSE 端点 URL。客户端应用此 URL 接收实时内容。"
          },
          "author": {
            "$ref": "#/components/schemas/User",
            "description": "Deprecated. 已弃用，保留兼容。"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "edited_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "attachment_count": {
            "type": "integer"
          },
          "mention_count": {
            "type": "integer"
          },
          "reaction_count": {
            "type": "integer"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reaction"
            }
          },
          "mentions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SendMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "type": {
            "type": "string",
            "description": "消息类型。\"stream\" 表示 AI 流式消息，将返回 SSE 而非 JSON。"
          },
          "source": {
            "$ref": "#/components/schemas/AISource",
            "description": "type=stream 时必须提供：AI 上游的 endpoint、auth_key 和 body"
          },
          "msg_id": {
            "type": "string",
            "description": "type=stream 时可选的消息 ID，不传则服务端生成"
          }
        }
      },
      "EditMessageRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          }
        }
      },
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "message_id",
          "filename",
          "mime_type",
          "size",
          "url"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message_id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "Reaction": {
        "type": "object",
        "required": [
          "emoji",
          "count",
          "me"
        ],
        "properties": {
          "emoji": {
            "type": "string",
            "example": "👍"
          },
          "count": {
            "type": "integer"
          },
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "选择该 reaction 的用户 ID 列表（ListReactions 时返回）"
          },
          "me": {
            "type": "boolean",
            "description": "当前用户是否选中该 reaction"
          }
        }
      }
    }
  }
}